==> 04_VRD_Competition_SubmissionFormat.ipynb <==
import pandas as pd
import gc
from collections import defaultdict
from tqdm import tqdm_notebook as tqdm
gc.collect()

sub_rel_dir = "submission_rel2.csv"
sub_is_dir = "submission_is.csv"

df_rel = pd.read_csv(sub_rel_dir)
df_is = pd.read_csv(sub_is_dir)
df_bbox_submission = pd.read_csv("submission.csv")
list_test_imageIds = df_bbox_submission["ImageId"].values
del df_bbox_submission

df = pd.concat([df_rel, df_is])
# df = df_rel

len(df)

list_imgids = list(df['ImageID'].unique())

len(list_imgids)

dict_pred = defaultdict(str)

for index, row in tqdm(df.iterrows()):
    imgid = row["ImageID"]
    confidence = (row["Confidence1"] + row["Confidence2"])/2
    label1 = row["LabelName1"]
    XMin1, YMin1, XMax1, YMax1 = row["XMin1"], row["YMin1"], row["XMax1"], row["YMax1"] 
    label2 = row["LabelName2"]
    XMin2, YMin2, XMax2, YMax2 = row["XMin2"], row["YMin2"], row["XMax2"], row["YMax2"] 
    relLabel = row["RelationshipLabel"]
    predStr = str(confidence) + " " + str(label1) + " " + \
      str(XMin1) + " "+ str(YMin1) + " " + str(XMax1) + " "+ str(YMax1) + " " + \
      str(label2) + " " + \
      str(XMin2) + " "+ str(YMin2) + " " + str(XMax2) + " "+ str(YMax2) + " " + \
      str(relLabel) + " "
    dict_pred[imgid]+=predStr

final_sub = []
for key, value in tqdm(dict_pred.items()):
    dict_res = {
      "ImageId": key,
      "PredictionString": value,
    }
    final_sub.append(dict_res)

df_submit = pd.DataFrame(final_sub)

list_predicted_imgs = df_submit.ImageId.values
main_list = list(set(list_test_imageIds) - set(list_predicted_imgs))
len(main_list)

append_unpredicted = []
for ml in main_list:
    append_unpredicted.append({
      "ImageId":ml,
      "PredictionString":""
    })
df_append = pd.DataFrame(append_unpredicted)

df_final = pd.concat([df_submit, df_append], ignore_index=True, sort =False)

df_final.head()

len(df_final)

df_final.to_csv("vrd_submission3_combined.csv", index=False)
